WIP net graph
