+++ runners/s3-benchrunner-rust/graph/http.py
@@ -1,0 +1,5 @@
+import plotly
+import plotly.graph_objs
+
+def draw(data) -> plotly.graph_objs.Figure:
+    pass
